Feat: rendering 함수 분리
