+++ index.html
@@ -109,5 +109,7 @@
       </main>
     </div>
     <script type="module" src="./src/router.js"></script>
+    <script src="./src/editor.js"></script>
+    <script src="./src/sidebar.js"></script>
   </body>
 </html>
